Rename allocator module
